pyproject.toml
[project]
name = "mashell"
version = "0.1.5"
description = "AI-powered command line assistant - any model, anywhere"
readme = "README.md"
license = {text = "GPL-3.0"}
requires-python = ">=3.11"
authors = [
    {name = "Jacob"}
]
keywords = ["ai", "cli", "llm", "shell", "assistant", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "prompt-toolkit>=3.0",
    "crawl4ai>=0.8.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.0.0",
    "slack-bolt>=1.18.0",
    "slack-sdk>=3.21.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "mypy>=1.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
mashell = "mashell.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = false
disallow_untyped_defs = false

# Allow subclasses to have more specific execute() signatures
# This is intentional design - tools define their own parameter types
[[tool.mypy.overrides]]
module = "mashell.tools.*"
disable_error_code = ["override"]
